fix: muda forma de executar
